feature[product] add product to cart and api relate

+++ src/main/java/org/graduate/shoefastbe/dto/attribute/AttributeDtoResponse.java
@@ -15,8 +15,8 @@ public class AttributeDtoResponse {
     private Double price;
     private Long productId;
     private Long size;
-    private Integer stock;
-    private Integer cache;
+    private Long stock;
+    private Long cache;
     private LocalDate createDate;
     private LocalDate modifyDate;
 }
